refactor: replace `moment` with `luxon`
