+++ src/utils/events/index.js
@@ -1,0 +1,14 @@
+/*
+
+	ard-eventhub
+	by SWR Audio Lab
+
+*/
+
+const createNewTopic = require('./createNewTopic')
+const processServices = require('./processServices')
+
+module.exports = {
+	createNewTopic,
+	processServices,
+}
